for merges, another chance

+++ Server/NEW/DATA/WebApi/Controllers/UserController.cs
@@ -6,6 +6,7 @@ using System.Net.Http;
 using System.Web.Http;
 using System.Web.Http.Cors;
 using DATA;
+using WebApi.DTO;
 
 namespace WebApi.Controllers
 {
@@ -14,7 +15,6 @@ namespace WebApi.Controllers
     {
         igroup194_Model db = new igroup194_Model();
 
-        // GET api/<controller>
         [HttpGet]
         [Route("GetUser/{id}")]
         public IHttpActionResult GetUser(int id)
@@ -22,7 +22,7 @@ namespace WebApi.Controllers
             try
             {
                 var user = db.tblUser.Where(x => x.Id == id).FirstOrDefault();
-                return Ok(user.FirstName + user.LastName);
+                return Ok(user.FirstName + " " + user.LastName);
             }
             catch (Exception ex)
             {
@@ -30,14 +30,20 @@ namespace WebApi.Controllers
             }
         }
 
+        
         //this is get method that return the user by email
         [HttpGet]
         [Route("GetUserEmail")]
-        public IHttpActionResult GetUserEmail([FromBody] string email)
+        public IHttpActionResult GetUserEmail([FromBody] UserDTO userEmail)
         {
+
             try
             {
-                var user = db.tblUser.Where(x => x.Email == email).FirstOrDefault();
+                var user = db.tblUser.Where(x => x.Email == userEmail.Email).FirstOrDefault();
+                if (user == null)
+                {
+                    return NotFound();
+                }
                 return Ok(user.Email);
             }
             catch (Exception ex)
@@ -46,13 +52,14 @@ namespace WebApi.Controllers
             }
         }
 
+        
         // insert user to db by calling Stored Prodecdure InsertUser
         [HttpPost]
         [Route("InsertUser")]
         public IHttpActionResult InsertUser([FromBody] tblUser user)
         {
             try
-             {
+            {
                 db.InsertUser(user.Email, user.Password, user.FirstName, user.LastName, user.gender, user.phoneNum, user.userUri);
                 db.SaveChanges();
                 return Ok("User Inserted Successfully");
@@ -63,20 +70,44 @@ namespace WebApi.Controllers
             }
         }
 
-        // PUT api/<controller>/5
-        public void Put(int id, [FromBody] string value)
+        
+        // update user to db by calling Function UpdateUser
+        [HttpPut]
+        [Route("UpdateUser")]
+        public IHttpActionResult UpdateUser([FromBody] UserDTO userToUpdate)
         {
+            try
+            {
+                tblUser user = db.tblUser.Where(x => x.Email == userToUpdate.Email).FirstOrDefault();
+                user.phoneNum = userToUpdate.phoneNum;
+                user.FirstName = userToUpdate.FirstName;
+                user.LastName = userToUpdate.LastName;
+                user.gender = userToUpdate.gender;
+                user.userUri = userToUpdate.userUri;
+                user.Password = userToUpdate.Password;
+                db.SaveChanges();
+                return Ok("User Updated Successfully");
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
         }
-
+        
+        
         //Only If user realy sure he wants to delete his account we will
         //delete him from the db - but first we will alert him to fuck himself
         [HttpDelete]
         [Route("DeleteUser")]
-        public IHttpActionResult Delete([FromBody] string email)
+        public IHttpActionResult Delete([FromBody] UserDTO userToDelete)
         {
             try
             {
-                var user = db.tblUser.Where(x => x.Email == email).FirstOrDefault();
+                var user = db.tblUser.Where(x => x.Email == userToDelete.Email).FirstOrDefault();
+                if (user == null)
+                {
+                    return NotFound();
+                }
                 db.tblUser.Remove(user);
                 db.SaveChanges();
                 return Ok("User Deleted Successfully");
